pyproject.toml
[project]
name = "class-practical-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.33",
    "openai>=1.109.1",
    "pymongo>=4.15.3",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
